requirements.txt
requests==2.31.0
pytest==7.4.0
allure-pytest==2.13.0
pytest-mock==3.12.0
pytest-timeout==2.2.0
Faker==19.13.0
pytest-cov==4.1.0